+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2024-09-09
+
+### Added
+- `PortDiff`s now have a `value` field to associate an integer value with the diff.
+- Add `PortDiffGraph::map_value` to set the `value` field of diffs. This creates new diffs with the same graph structure.
+
+
 ## [0.3.0] - 2024-09-04
 
 ### Added